refactor: move mocks under tests
